# app-annonces/src/swagger.yml
openapi: 3.0.0
info:
  title: API Plateforme d'Annonces
  license:
    name: ISC
servers:
  - url: http://localhost:3000/api
    description: Serveur de développement
  - url: https://api.annonces.com/api
    description: Serveur de production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu après connexion
  schemas:
...
    Signalement:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "user@example.com"
        message:
          type: string
          example: "Cette annonce contient des informations incorrectes"
        annonce_id:
          type: integer
          example: 5
        annonce_titre:
          type: string
          example: "MacBook Pro 2020"
        statut:
          type: string
          enum: [nouveau, traite, rejete]
          example: "nouveau"
        created_at:
          type: string
          format: date-time
...
paths:
...
  /signalements:
    post:
      tags:
        - Signalements
      summary: Signaler une annonce
      description: Créer un signalement pour une annonce problématique
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - message
                - annonce_id
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                message:
                  type: string
                  minLength: 10
                  maxLength: 1000
                  example: "Cette annonce contient des informations incorrectes"
                annonce_id:
                  type: integer
                  minimum: 1
                  example: 5
      responses:
        '201':
          description: Signalement créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Signalement enregistré avec succès. L'administrateur sera notifié."
                  signalementId:
                    type: integer
                    example: 1
        '400':
          description: Erreur de validation
        '404':
          description: Annonce non trouvée

    get:
      tags:
        - Signalements
      summary: Liste des signalements (Admin)
      description: Récupérer tous les signalements (administrateur uniquement)
      security:
        - BearerAuth: []
      parameters:
        - name: statut
          in: query
          description: Filtrer par statut
          required: false
          schema:
            type: string
            enum: [nouveau, traite, rejete]
            example: "nouveau"
        - name: annonce_id
          in: query
          description: Filtrer par annonce
          required: false
          schema:
            type: integer
            example: 5
      responses:
        '200':
          description: Liste des signalements
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Liste des signalements"
                  count:
                    type: integer
                    example: 3
                  signalements:
                    type: array
                    items:
                      $ref: '#/components/schemas/Signalement'
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé (non administrateur)
  /signalements/{id}/statut:
    patch:
      tags:
        - Signalements
      summary: Modifier statut signalement (Admin)
      description: Changer le statut d'un signalement (administrateur uniquement)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID du signalement
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - statut
              properties:
                statut:
                  type: string
                  enum: [nouveau, traite, rejete]
                  example: "traite"
      responses:
        '200':
          description: Statut modifié
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Statut du signalement mis à jour avec succès"
                  new_statut:
                    type: string
                    example: "traite"
        '404':
          description: Signalement non trouvé
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé
...